260719 - Removed models folder and moved the interfaces and schema folders to src folder
