+++ src/schema/environmentConfig.schema.ts
